seperate KeyFramer class from animation class
